+++ packages/atom.io/mutable-cache/package.json
@@ -1,0 +1,15 @@
+{
+	"name": "atom.io-mutable-cache",
+	"private": true,
+	"main": "dist/index.js",
+	"types": "dist/index.d.ts",
+	"module": "dist/index.mjs",
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"browser": "./dist/index.mjs",
+			"import": "./dist/index.mjs",
+			"require": "./dist/index.js"
+		}
+	}
+}
