+++ packages/atom.io/src/internal/index.ts
@@ -7,7 +7,7 @@ export * from "./selector-internal"
 export * from "./selector"
 export * from "./set"
 export * from "./store"
-export * from "./subject"
+export * from "../../../anvl/reactivity/subject"
 export * from "./subscribe-internal"
 export * from "./time-travel-internal"
 export * from "./timeline-internal"
