+++ packages/atom.io/tsconfig.json
@@ -5,5 +5,5 @@
 		"jsxImportSource": "react",
 		"types": ["vitest/globals"]
 	},
-	"include": ["./src/**/*", "./test/**/*"]
+	"include": ["./src/**/*", "./test/**/*", "../anvl/reactivity/subject.ts"]
 }
